+++ src/main/resources/application.properties
@@ -2,10 +2,10 @@
 server.port=8081
 
 # DataSource 관련 설정.
-spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
-spring.datasource.url=jdbc:oracle:thin:@localhost:1521:xe
-spring.datasource.username=jpastudy
-spring.datasource.password=jpastudy
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/testDb
+spring.datasource.username=root
+spring.datasource.password=123123
 
 # JPA/Hibernate에서 DDL(create, drop, alter) 문장을 자동으로 작성/실행.
 # 스프링 부트 어플리케이션이 실행될 때 drop table을 실행하고 create table을 실행.
